scraper/requirements.txt
# The Hive Instagram Scraper
# Uses Playwright to scrape Instagram posts from ITU club pages

playwright>=1.40.0
requests>=2.31.0
python-dateutil>=2.8.2
openai>=1.0.0
python-dotenv>=1.0.0
